+++ SRC/Model/DAO/AutoTipusDAO.php
@@ -1,10 +1,10 @@
 <?php
 $path = dirname(__DIR__, 2);
 include($path . "/Connection/Dbconn.php");
-function AutoTipusTarolo($Marka,$Tipus ,$Fajta, $Kategoria, $Premium, $KornyezetvedelmiBesorolas,$Ar)
+function AutoTipusTarolo($Marka,$Tipus,$Premium,$Ar)
 {
     $kapcsolat = $GLOBALS["kapcsolat"];
-    $sql = "INSERT INTO `autokolcsonzo`.`autotipus` (`ID`, `Márka`,`Tipus`, `Fajta_ID`, `Kategoria_ID`, `Prémium`, `Környezetvédelmi_ID`, `Ar`) VALUES (NULL, '$Marka','$Tipus', '$Fajta', '$Kategoria', '$Premium', '$KornyezetvedelmiBesorolas','$Ar')"; // modositott
+    $sql = "INSERT INTO `autokolcsonzo`.`autotipus` (`ID`, `Márka`,`Tipus`,  `Prémium`, `Ar`) VALUES (NULL, '$Marka','$Tipus', '$Premium', '$Ar')"; // modositott
 
     $üzenet = "Az autotipus felvétel ";
     $ok = mysqli_query($kapcsolat, $sql);
@@ -19,25 +19,4 @@ function SQLFeltolto($kapcsolat, $sql, $value)
         $SQLTomb[$egysor["ID"]] = $egysor[$value];
     }
     return $SQLTomb;
-}
-function FajtaFeltolto($kapcsolat)
-{
-    $sql = "SELECT * FROM `autokolcsonzo`.`fajta`";
-    $Fajta = [];
-    $Fajta = SQLFeltolto($kapcsolat, $sql, "Fajta_neve");
-    return $Fajta;
-}
-function KategoriaFeltolto($kapcsolat)
-{
-    $sql = "SELECT * FROM `autokolcsonzo`.`kategoria`";
-    $Kategoria = [];
-    $Kategoria = SQLFeltolto($kapcsolat, $sql, "Kategoria");
-    return $Kategoria;
-}
-function KornyezetVedelemFeltolto($kapcsolat)
-{
-    $sql = "SELECT * FROM `autokolcsonzo`.`környezetvédelmibesorolás`";
-    $Környezetvédelmibesorolás = [];
-    $Környezetvédelmibesorolás = SQLFeltolto($kapcsolat, $sql, "KörnyezetvédelmiBesorolás");
-    return $Környezetvédelmibesorolás;
 }
