--- SRC/Model/Service/ArService.php
@@ -1,8 +1,0 @@
-<?php
-$path = dirname(__DIR__, 1);
-include_once($path."/DAO/ArDAO.php");
-function AutoArAtvevo($markatipus,$ar){
-    return AutoarTarolo($markatipus,$ar); //ar beszur'as
- }
-
-?>
